Added more timelines and dr overlap and defensive overlaps

+++ src/main/java/com/samtius/arenaparser/dto/MatchCombatDetails.java
@@ -98,7 +98,7 @@ public record MatchCombatDetails(
     }
 
     public record TimelineEvent(
-            int offsetSeconds,
+            double offsetSeconds,
             long spellId,
             String spell,
             String category,
